// src/types/Color.ts
export type Color =
  /* Kept default Bootstrap's palette color */
  | "light"
  | "muted"
  | "dark"
  | "gold"
  | "light-gold"
  | "blue"
  | "light-blue"
  | "deep"
  | "light-deep"
  | "jewel"
  | "skin"
  | "dark-skin"
  | "light-skin"
  | "green"
  | "light-green"
  | "dark-green"
  | "purple"
  | "light-purple"
  | "dark-purple"
  | "brown"
  | "light-brown"
  | "turquoize"
  | "dark-turquoize"
  | "orange";